requirements.txt
dvc[gs]==2.37.0
pytorch-fid==0.3.0
accelerate
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
